changed test file to be similar in format to test_nirs
